EditEventPage is now working as a base for clone event page.
PT-47
