+++ src/domain/app/i18n/sv.json
@@ -557,6 +557,7 @@
     "buttonBack": "Evenemang",
     "buttonCreateOccurrence": "Lägg till en ny evenemangtid",
     "buttonEventDetails": "Visa information om evenemanget",
+    "buttonCloneEventDetails": "Kopiera till en ny mall",
     "buttonShowMore": "Visa fler händelser",
     "count": "{{count}} st",
     "deleteError": "Det gick inte att ta bort evenemangtiden",
