+++ src/domain/event/EditEventPage.tsx
@@ -1,3 +1,4 @@
+import { compact } from 'lodash';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory, useLocation, useParams } from 'react-router';
@@ -8,7 +9,6 @@ import {
   PersonFieldsFragment,
   useEditEventMutation,
   useEventQuery,
-  useUpdateSingleImageMutation,
 } from '../../generated/graphql';
 import useLocale from '../../hooks/useLocale';
 import { useSearchParams } from '../../hooks/useQuery';
@@ -19,19 +19,19 @@ import PageWrapper from '../app/layout/PageWrapper';
 import { ROUTES } from '../app/routes/constants';
 import ErrorPage from '../errorPage/ErrorPage';
 import { PUBLICATION_STATUS } from '../events/constants';
-import { getImageName } from '../image/utils';
 import ActiveOrganisationInfo from '../organisation/activeOrganisationInfo/ActiveOrganisationInfo';
-import { createOrUpdateVenue } from '../venue/utils';
-import { VIRTUAL_EVENT_LOCATION_ID } from './constants';
 import EventForm, { defaultInitialValues } from './eventForm/EventForm';
+import {
+  useCreateOrUpdateVenueRequest,
+  useUpdateImageRequest,
+} from './eventForm/useEventFormSubmitRequests';
 import styles from './eventPage.module.scss';
 import { EventFormFields } from './types';
 import {
+  getEventFormValues,
   getEventLanguageFromUrl,
   getEventPayload,
-  getEventVenueDescription,
   getFirstAvailableLanguage,
-  getRealKeywords,
   isEditableEvent,
 } from './utils';
 
@@ -71,7 +71,6 @@ const EditEventPage: React.FC = () => {
     ) || [];
 
   const [editEvent] = useEditEventMutation();
-  const [updateImage] = useUpdateSingleImageMutation();
 
   const goToEventDetailsPage = () => {
     history.push(`/${locale}${ROUTES.EVENT_DETAILS.replace(':id', id)}`);
@@ -101,17 +100,27 @@ const EditEventPage: React.FC = () => {
     }
   }, [eventData, language]);
 
+  React.useEffect(() => {
+    if (eventData && selectedLanguage) {
+      setInitialValues(getEventFormValues(eventData, selectedLanguage));
+    }
+  }, [eventData, selectedLanguage]);
+
   const shouldSaveImage = (values: EventFormFields): boolean =>
     !!values.image &&
     (values.image !== initialValues.image ||
       values.imageAltText !== initialValues.imageAltText ||
       values.imagePhotographerName !== initialValues.imagePhotographerName);
 
+  const createOrUpdateVenueRequestHandler = useCreateOrUpdateVenueRequest(
+    selectedLanguage
+  );
+  const updateImageRequestHandler = useUpdateImageRequest();
+
   const submit = async (values: EventFormFields) => {
     try {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const requests: Promise<any>[] = [];
-      requests.push(
+      const requests: Promise<any>[] = compact([
         editEvent({
           variables: {
             event: {
@@ -130,36 +139,10 @@ const EditEventPage: React.FC = () => {
                 PUBLICATION_STATUS.DRAFT,
             },
           },
-        })
-      );
-
-      const createOrUpdateVenueRequest = createOrUpdateVenue({
-        venueFormData: values,
-        language: selectedLanguage,
-        locationId: values.location,
-      });
-
-      if (createOrUpdateVenueRequest) {
-        requests.push(createOrUpdateVenueRequest);
-      }
-
-      if (shouldSaveImage(values)) {
-        const imageName = getImageName(values.image);
-        if (imageName) {
-          requests.push(
-            updateImage({
-              variables: {
-                image: {
-                  altText: values.imageAltText,
-                  id: values.image,
-                  name: imageName,
-                  photographerName: values.imagePhotographerName,
-                },
-              },
-            })
-          );
-        }
-      }
+        }),
+        createOrUpdateVenueRequestHandler(values),
+        shouldSaveImage(values) ? updateImageRequestHandler(values) : undefined,
+      ]);
 
       // Run all requests parallel
       await Promise.all(requests);
@@ -185,82 +168,13 @@ const EditEventPage: React.FC = () => {
     setSelectedLanguage(newLanguage);
   };
 
-  const getLocationId = (locationId: string | undefined | null): string => {
-    // If location id is virtual event location, we are not going to show it as location.
-    // Only virtual location checkbox should be checked.
-    if (locationId !== VIRTUAL_EVENT_LOCATION_ID) {
-      return locationId ?? '';
-    }
-    return '';
-  };
-
-  React.useEffect(() => {
-    if (eventData) {
-      setInitialValues({
-        additionalCriteria:
-          eventData.event?.additionalCriteria.map((item) => item.id || '') ||
-          [],
-        categories:
-          eventData.event?.categories.map((item) => item.id || '') || [],
-        audience: eventData.event?.audience.map((item) => item.id || '') || [],
-        contactEmail: eventData.event?.pEvent?.contactEmail || '',
-        contactPersonId: eventData.event?.pEvent?.contactPerson?.id || '',
-        contactPhoneNumber: eventData.event?.pEvent?.contactPhoneNumber || '',
-        description: eventData.event?.description?.[selectedLanguage] || '',
-        enrolmentEndDays:
-          eventData.event?.pEvent?.enrolmentEndDays?.toString() || '',
-        enrolmentStart: eventData.event?.pEvent?.enrolmentStart
-          ? new Date(eventData.event?.pEvent?.enrolmentStart)
-          : null,
-        image: eventData.event?.images[0]?.id || '',
-        imageAltText: eventData.event?.images[0]?.altText || '',
-        imagePhotographerName:
-          eventData.event?.images[0]?.photographerName || '',
-        infoUrl: eventData.event?.infoUrl?.[selectedLanguage] || '',
-        inLanguage:
-          eventData.event?.inLanguage.map((item) => item.id || '') || [],
-        isFree: !!eventData.event?.offers?.[0]?.isFree,
-        priceDescription:
-          eventData.event?.offers?.[0]?.description?.[selectedLanguage] || '',
-        keywords:
-          getRealKeywords(eventData)?.map((keyword) => keyword.id || '') || [],
-        location: getLocationId(eventData.event?.location.id),
-        name: eventData.event?.name[selectedLanguage] || '',
-        neededOccurrences:
-          eventData.event?.pEvent?.neededOccurrences.toString() || '',
-        price: eventData.event?.offers?.[0]?.price?.[selectedLanguage] || '',
-        shortDescription:
-          eventData.event?.shortDescription?.[selectedLanguage] || '',
-        locationDescription: getEventVenueDescription(
-          eventData,
-          selectedLanguage
-        ),
-        hasClothingStorage:
-          eventData?.event?.venue?.hasClothingStorage || false,
-        hasSnackEatingPlace:
-          eventData?.event?.venue?.hasSnackEatingPlace || false,
-        outdoorActivity: eventData?.event?.venue?.outdoorActivity || false,
-        hasToiletNearby: eventData?.event?.venue?.hasToiletNearby || false,
-        hasAreaForGroupWork:
-          eventData?.event?.venue?.hasAreaForGroupWork || false,
-        hasIndoorPlayingArea:
-          eventData?.event?.venue?.hasIndoorPlayingArea || false,
-        hasOutdoorPlayingArea:
-          eventData?.event?.venue?.hasOutdoorPlayingArea || false,
-        autoAcceptance: eventData.event?.pEvent.autoAcceptance,
-        mandatoryAdditionalInformation:
-          eventData.event?.pEvent?.mandatoryAdditionalInformation || false,
-        isVirtual: eventData.event?.location.id === VIRTUAL_EVENT_LOCATION_ID,
-      });
-    }
-  }, [eventData, selectedLanguage]);
-
+  const isCloningEvent = true;
   return (
     <PageWrapper title="editEvent.pageTitle">
       <LoadingSpinner isLoading={loading}>
         {!!eventData ? (
           <>
-            {isEditableEvent(eventData) ? (
+            {isEditableEvent(eventData) || isCloningEvent ? (
               <Container>
                 <div className={styles.eventPage}>
                   <ActiveOrganisationInfo organisationId={organisationId} />
